# gateway/internal/server/chat.go
package server

import (
	"bufio"
	"bytes"
	"context"
	"encoding/json"
	"errors"
	"fmt"
	"io"
	"log/slog"
	"math"
	"strconv"
	"sync/atomic"
	"time"

	"github.com/gofiber/fiber/v2"

	"github.com/cognigate/gateway/internal/apierr"
	"github.com/cognigate/gateway/internal/cache"
	"github.com/cognigate/gateway/internal/httpx"
	"github.com/cognigate/gateway/internal/obs"
	"github.com/cognigate/gateway/internal/provider"
	"github.com/cognigate/gateway/internal/routing"
	"github.com/cognigate/gateway/internal/store"
)

// chatEnvelope is the only part of a completion request the gateway parses.
//
// Everything else — messages, tools, provider extensions — is forwarded byte
// for byte. Reading a field here never removes it from what the upstream
// receives; the request is relayed from the original bytes, and this struct is
// only how the gateway learns what it must decide.
//
// Routing needs the model and whether the caller wants a stream. The three
// sampling parameters are here for GW-12's eligibility rule and nothing else:
// they are numbers, not content, so reading them does not touch what GW-14 puts
// out of bounds.
type chatEnvelope struct {
	Model  string `json:"model"`
	Stream bool   `json:"stream"`
	// Raw rather than *float64/*int, because a typed decode makes the gateway
	// refuse requests the upstream would accept: {"n": 1.0} is what a Python
	// client's json.dumps of a float emits, and it does not fit an int. These
	// three are read only to decide cache eligibility, so a value the gateway
	// cannot read is answered by declining to cache it, never by rejecting it.
	Temperature json.RawMessage `json:"temperature"`
	TopP        json.RawMessage `json:"top_p"`
	N           json.RawMessage `json:"n"`
}

// notNumber is what a present but non-numeric sampling field reads as. NaN
// compares false against everything, so such a request fails every eligibility
// test without needing a branch of its own.
var notNumber = math.NaN()

// sampled reads one of the three numeric fields. Absent and null both read as
// absent, which is the same question GW-12 asks: did the caller express this?
func sampled(raw json.RawMessage) (value float64, present bool) {
	if len(raw) == 0 || string(raw) == "null" {
		return 0, false
	}
	var f float64
	if err := json.Unmarshal(raw, &f); err != nil {
		return notNumber, true
	}
	return f, true
}

// deterministic reports whether the request, as written, asks for randomness.
//
// This is GW-12's rule verbatim: temperature 0, or no temperature alongside the
// API defaults for top_p and n. Note what it does not do — a request that omits
// temperature entirely is treated as deterministic even though the provider
// will sample at its own default. That is deliberate in the spec ("deterministic
// as expressed") and safe only because caching is opt-in: nothing is cached for
// a caller who did not ask for it.
func (e chatEnvelope) deterministic() bool {
	if t, present := sampled(e.Temperature); present {
		return t == 0
	}
	p, pSet := sampled(e.TopP)
	n, nSet := sampled(e.N)
	return (!pSet || p == 1) && (!nSet || n == 1)
}

func (s *Server) handleChatCompletions(c *fiber.Ctx) error {
	body := c.Body()
	if len(body) == 0 {
		return httpx.Fail(c, apierr.InvalidRequest("A JSON request body is required."))
	}

	var env chatEnvelope
	if err := json.Unmarshal(body, &env); err != nil {
		return httpx.Fail(c, apierr.InvalidRequest("Request body is not valid JSON.").WithCause(err))
	}
	if env.Model == "" {
		return httpx.Fail(c, apierr.InvalidRequest("The \"model\" field is required.").WithParam("model"))
	}

	tenantID := httpx.TenantID(c)

	quota, err := s.evaluateQuota(c.UserContext(), requestScope(c))
	if err != nil {
		return httpx.Fail(c, err)
	}
	c.Set(httpx.HeaderQuotaState, quota.State)
	if quota.Reject != nil {
		s.recordRejection(c, env.Model, quota.Reject)
		return httpx.Fail(c, quota.Reject)
	}

	// The body is copied because fasthttp reuses its buffer once the handler
	// returns, and a streamed response outlives the handler.
	payload := append([]byte(nil), body...)

	plan := s.planCache(c, env)
	if plan.header != "" {
		setCacheStatus(c, plan.header)
	}

	if env.Stream {
		return s.streamCompletion(c, tenantID, env.Model, payload)
	}
	return s.completion(c, tenantID, env.Model, payload, plan)
}

// completion serves a buffered chat completion, from the cache when GW-12 says
// it may and the answer is there.
func (s *Server) completion(c *fiber.Ctx, tenantID, requested string, body []byte, plan cachePlan) error {
	ctx, cancel := s.opContext(c)
	defer cancel()

	started := time.Now()

	// servedBy is the model the key was computed against, kept so that storing
	// can check the answer came from it. Empty means this request is not
	// cacheable, and both the lookup and the store are then skipped.
	key, servedBy := s.cacheKeyFor(ctx, plan, tenantID, requested, body)
	if key != "" {
		if entry, ok := s.cache.Get(key); ok {
			return s.serveFromCache(c, entry, requested, started)
		}
	}

	result, err := s.Dispatcher.Dispatch(ctx, tenantID, requested, body, false, "/chat/completions")
	if err != nil {
		s.recordRejection(c, requested, err)
		return httpx.Fail(c, timeoutAware(ctx, s.limits(c).RequestTimeout, err))
	}
	defer result.Response.Close()

	s.applyRoutingHeaders(c, result)
	s.meter(c, result, requested, result.Response.Usage, false, fiber.StatusOK, time.Since(started))

	if key != "" {
		setCacheStatus(c, cacheMiss)
		// Stored only when the model that answered is the one the key names. A
		// cascade that fell through to a second model produced a different
		// answer than the key promises, and replaying it later as the primary's
		// would make the fallback invisible to the caller it happened to.
		if result.Candidate.ServedBy() == servedBy {
			s.cache.Put(tenantID, key, cache.Entry{
				Body:     append([]byte(nil), result.Response.Body...),
				Provider: result.Candidate.Provider,
				Model:    result.Candidate.Model,
			}, plan.ttl)
		}
	}

	c.Set(fiber.HeaderContentType, fiber.MIMEApplicationJSON)
	return c.Status(fiber.StatusOK).Send(result.Response.Body)
}

// cacheKeyFor derives this request's cache key, or "" when it has none.
//
// Resolution runs here rather than being taken from the dispatch, because the
// key names the model that will serve and a lookup that waited for the dispatch
// would have nothing left to save. A resolution or parse failure yields no key
// rather than an error: the dispatch a few lines below fails on the same input
// and says so properly, and reporting it from the cache would blame the wrong
// component.
func (s *Server) cacheKeyFor(ctx context.Context, plan cachePlan, tenantID, requested string, body []byte) (key, servedBy string) {
	if !plan.lookup {
		return "", ""
	}
	primary, err := s.Dispatcher.Primary(ctx, tenantID, requested)
	if err != nil {
		return "", ""
	}
	key, err = cache.Key(tenantID, primary.ServedBy(), body)
	if err != nil {
		return "", ""
	}
	return key, primary.ServedBy()
}

// serveFromCache replays a stored answer.
//
// The body goes back byte for byte, its original completion id and usage block
// included, because a caller that retried an identical request is entitled to an
// identical answer — that is the whole promise. Only the headers differ, and
// they differ honestly: the request id is this request's, and the cache header
// says where the body came from.
func (s *Server) serveFromCache(c *fiber.Ctx, entry cache.Entry, requested string, started time.Time) error {
	setCacheStatus(c, cacheHit)
	c.Set(httpx.HeaderServedBy, entry.ServedBy())
	c.Set(httpx.HeaderFallbackDepth, "0")

	out := httpx.GetOutcome(c)
	out.Provider = entry.Provider
	out.Model = entry.Model
	if requested != entry.Model {
		out.Alias = requested
	}
	httpx.SetOutcome(c, out)

	// No tokens, no cost, no upstream-latency observation: nothing was consumed
	// and nothing was called. The row exists so the request is countable and so
	// a tenant can see how much of its traffic the cache absorbed (GW-12.AC-2).
	if s.Telemetry != nil {
		s.Telemetry.Record(store.UsageRecord{
			RequestID:       httpx.RequestID(c),
			ClientRequestID: httpx.ClientRequestID(c),
			TenantID:        httpx.TenantID(c),
			KeyPrefix:       keyPrefix(c),
			Provider:        entry.Provider,
			Model:           entry.Model,
			RequestedModel:  requested,
			Cached:          true,
			StatusCode:      fiber.StatusOK,
			DurationMS:      time.Since(started).Milliseconds(),
			RecordedAt:      time.Now().UTC(),
		})
	}

	c.Set(fiber.HeaderContentType, fiber.MIMEApplicationJSON)
	return c.Status(fiber.StatusOK).Send(entry.Body)
}

// streamCompletion relays an SSE completion.
//
// The upstream body is handed to Fiber's stream writer, which runs after this
// handler returns — so ownership of Close moves into the writer. Deferring it
// here would close the upstream before a single token had been relayed.
func (s *Server) streamCompletion(c *fiber.Ctx, tenantID, requested string, body []byte) error {
	// No request deadline on a stream: it is bounded by silence
	// (stream_idle_timeout), not by duration.
	ctx := c.UserContext()

	started := time.Now()
	result, err := s.Dispatcher.Dispatch(ctx, tenantID, requested, body, true, "/chat/completions")
	if err != nil {
		// Nothing has been written yet, so this is still an ordinary HTTP error
		// with a status line. GW-3's "fall back only before the first byte" rule
		// is satisfied by construction: the cascade lives entirely above here.
		s.recordRejection(c, requested, err)
		return httpx.Fail(c, timeoutAware(ctx, s.limits(c).RequestTimeout, err))
	}

	s.applyRoutingHeaders(c, result)
	s.reportOutcome(c, result, requested, nil)
	c.Set(fiber.HeaderContentType, "text/event-stream")
	c.Set(fiber.HeaderCacheControl, "no-cache")
	c.Set(fiber.HeaderConnection, "keep-alive")
	// Proxies that buffer will hold the whole completion and defeat streaming;
	// this is the header nginx and its imitators honour.
	c.Set("X-Accel-Buffering", "no")
	c.Status(fiber.StatusOK)

	// Everything the usage record needs is read here, while the request is still
	// alive, and carried into the writer by the closure. Reading any of it from
	// the Fiber context inside the writer would be reading a recycled request:
	// the handler has returned by the time the writer runs.
	var (
		requestID       = httpx.RequestID(c)
		clientRequestID = httpx.ClientRequestID(c)
		prefix          = keyPrefix(c)
		logger          = s.Logger
		// The tenant's own stall budget, not the deployment's. Read here for the
		// same reason as everything else in this block: by the time the writer
		// runs, c is a recycled request and the tenant on it belongs to someone
		// else.
		idle = s.limits(c).StreamIdleTimeout
		// The concurrency permit moves with the relay. Taken here rather than in
		// the writer for the same reason as everything else in this block: c is a
		// recycled request by then.
		releaseSlot = adoptSlot(c)
	)

	c.Context().SetBodyStreamWriter(func(w *bufio.Writer) {
		defer releaseSlot()
		defer result.Response.Close()

		usage, stalled, relayErr := relaySSE(w, result.Response.Stream, idle)

		switch {
		case stalled:
			// The status line went out long ago, so the only way to report this
			// is in-band. GW-7 gives the stall its own code precisely because it
			// can never be an HTTP status.
			writeStreamError(w, apierr.UpstreamError("The upstream stopped sending data.").
				WithCause(relayErr), apierr.CodeUpstreamStreamStalled, requestID)
			logger.Warn("upstream stream stalled",
				slog.String("request_id", requestID),
				slog.String("served_by", result.Candidate.ServedBy()))
		case relayErr != nil && !errors.Is(relayErr, io.EOF):
			// The upstream died part-way through — a dropped connection, a reset,
			// a provider that panicked. Without a terminal event this is
			// indistinguishable from a completion that simply ended, and GW-3.AC-7
			// forbids leaving the caller to guess: a truncated answer must announce
			// itself. If the reason the relay stopped is that the *client* hung up,
			// this frame goes to a closed socket and is discarded, which costs
			// nothing and is not worth a special case.
			writeStreamError(w, apierr.UpstreamError("The upstream ended the stream before it completed.").
				WithCause(relayErr), apierr.CodeUpstreamError, requestID)
			logger.Warn("stream relay ended early",
				slog.String("request_id", requestID),
				slog.String("served_by", result.Candidate.ServedBy()),
				slog.String("error", relayErr.Error()))
		}

		s.meterStream(requestID, clientRequestID, tenantID, prefix, result, requested, usage, time.Since(started))
	})

	return nil
}

// relaySSE copies an event stream through, watching for silence.
//
// It returns the usage block if the upstream sent one, whether the stream
// stalled, and the error that ended the relay.
func relaySSE(w *bufio.Writer, src io.ReadCloser, idleTimeout time.Duration) (*provider.Usage, bool, error) {
	if idleTimeout <= 0 {
		idleTimeout = 60 * time.Second
	}

	var stalled atomic.Bool
	// Closing the body is what unblocks the read: there is no way to interrupt
	// an in-flight Read otherwise, and a goroutine parked on a dead connection
	// would leak for as long as the process lives.
	watchdog := time.AfterFunc(idleTimeout, func() {
		stalled.Store(true)
		_ = src.Close()
	})
	defer watchdog.Stop()

	var usage *provider.Usage
	reader := bufio.NewReaderSize(src, 16*1024)

	for {
		line, err := reader.ReadBytes('\n')
		if len(line) > 0 {
			watchdog.Reset(idleTimeout)

			if u := usageFromEvent(line); u != nil {
				usage = u
			}
			if _, werr := w.Write(line); werr != nil {
				// The client hung up. Nothing to report to anyone but the log.
				return usage, false, werr
			}
			// Flushing per line is the point of a stream; buffering would
			// reintroduce exactly the latency the caller asked to avoid.
			if werr := w.Flush(); werr != nil {
				return usage, false, werr
			}
		}
		if err != nil {
			if stalled.Load() {
				return usage, true, err
			}
			if errors.Is(err, io.EOF) {
				return usage, false, nil
			}
			return usage, false, err
		}
	}
}

// sseUsage is the only field read out of a streamed chunk. Decoding into this
// shape rather than the full chunk means completion text is never materialised
// into a Go value on the metering path (GW-14).
type sseUsage struct {
	Usage *provider.Usage `json:"usage"`
}

func usageFromEvent(line []byte) *provider.Usage {
	const prefix = "data: "
	trimmed := bytes.TrimSpace(line)
	if !bytes.HasPrefix(trimmed, []byte(prefix)) {
		return nil
	}
	payload := bytes.TrimSpace(trimmed[len(prefix):])
	if len(payload) == 0 || bytes.Equal(payload, []byte("[DONE]")) {
		return nil
	}
	var parsed sseUsage
	if json.Unmarshal(payload, &parsed) != nil {
		return nil
	}
	if parsed.Usage == nil || parsed.Usage.TotalTokens == 0 {
		return nil
	}
	return parsed.Usage
}

// writeStreamError emits a terminal error event in the GW-7 envelope, then the
// [DONE] sentinel so a well-behaved client's loop still terminates.
func writeStreamError(w *bufio.Writer, e *apierr.Error, code, requestID string) {
	envelope := e.Envelope(requestID)
	envelope.Error.Code = code
	payload, err := json.Marshal(envelope)
	if err != nil {
		return
	}
	fmt.Fprintf(w, "event: error\ndata: %s\n\n", payload)
	fmt.Fprint(w, "data: [DONE]\n\n")
	_ = w.Flush()
}

// --- response headers and metering ------------------------------------------

// applyRoutingHeaders sets the extension headers that describe how a successful
// request was actually served.
func (s *Server) applyRoutingHeaders(c *fiber.Ctx, result *routing.Result) {
	c.Set(httpx.HeaderServedBy, result.Candidate.ServedBy())
	c.Set(httpx.HeaderFallbackDepth, strconv.Itoa(result.Depth))
	if result.Candidate.Entry.Deprecated {
		c.Set(httpx.HeaderDeprecation, "true")
	}
}

// reportOutcome hands the request-scoped facts to the observability middleware,
// which sees only the route and the status and would otherwise log a completion
// without saying what completed it.
//
// It must be called while the request is alive. On the streaming path that means
// before the writer starts, which is why a stream reports no token counts: they
// are not known until the upstream body has been relayed, and by then this
// context has been recycled. The usage record, which is written from the writer
// goroutine and carries its own copies, is where a stream's tokens are found.
func (s *Server) reportOutcome(c *fiber.Ctx, result *routing.Result, requested string, usage *provider.Usage) {
	out := httpx.GetOutcome(c)
	out.Provider = result.Candidate.Provider
	out.Model = result.Candidate.Model
	out.FallbackDepth = result.Depth
	out.UpstreamMS = result.Response.Latency.Milliseconds()
	// Only an alias, never the resolved id: "requested" equal to the served
	// model means the caller named the model directly, and repeating it in a
	// field named alias would read as an alias that happens to share its name.
	if requested != result.Candidate.Model {
		out.Alias = requested
	}
	if usage != nil {
		out.PromptTokens = usage.PromptTokens
		out.CompletionTokens = usage.CompletionTokens
	}
	httpx.SetOutcome(c, out)
}

// meterCascade counts the hops a request took to reach the candidate that
// served it, one row per hop rather than one per request.
//
// Per-hop is what makes the series answerable: "the chain from A ended at C"
// says nothing about which link failed, whereas A→B and B→C say that A is the
// one that broke. The reason comes from the failure enumeration, never from
// upstream error text — that is unbounded as a label value, and under GW-14 it
// may quote the request that produced it.
func (s *Server) meterCascade(tenantID string, result *routing.Result) {
	if result.Depth == 0 || len(result.Attempts) < 2 {
		return
	}
	for i := 1; i < len(result.Attempts); i++ {
		from, to := result.Attempts[i-1], result.Attempts[i]
		s.Metrics.FallbackCascades.WithLabelValues(
			tenantID, from.Candidate.Model, to.Candidate.Model, cascadeReason(from),
		).Inc()
	}
}

// cascadeReason names why one attempt handed off to the next, in the small
// vocabulary a dashboard can group by.
func cascadeReason(a routing.Attempt) string {
	if a.Skipped {
		return "breaker_open"
	}
	return a.Failure.String()
}

// meter records one served request: metrics now, usage row asynchronously.
func (s *Server) meter(c *fiber.Ctx, result *routing.Result, requested string, usage *provider.Usage, streamed bool, status int, elapsed time.Duration) {
	s.reportOutcome(c, result, requested, usage)
	s.record(httpx.RequestID(c), httpx.ClientRequestID(c), httpx.TenantID(c), keyPrefix(c),
		result, requested, usage, streamed, status, elapsed)
}

// meterStream is meter's counterpart for the streaming path, where the Fiber
// context's own buffers are no longer safe to read: the handler has returned and
// fasthttp may have recycled them. So the identity of the request is passed in
// as plain strings, captured by streamCompletion while the request was alive. A
// streamed completion is metered exactly like a buffered one — anything less and
// the most common shape of traffic would be missing from usage, quota and
// billing.
func (s *Server) meterStream(
	requestID, clientRequestID, tenantID, prefix string,
	result *routing.Result, requested string,
	usage *provider.Usage, elapsed time.Duration,
) {
	s.record(requestID, clientRequestID, tenantID, prefix,
		result, requested, usage, true, fiber.StatusOK, elapsed)
}

func (s *Server) record(
	requestID, clientRequestID, tenantID, prefix string,
	result *routing.Result, requested string,
	usage *provider.Usage, streamed bool, status int, elapsed time.Duration,
) {
	cand := result.Candidate
	cost := result.CostUSD(usage)
	// What the call cost the operator and what its tenant owes for it are
	// recorded separately, so the margin between them is a figure rather
	// than an assumption. They are equal under the default mode.
	charge := s.Config.Billing.Charge(cost)

	if s.Metrics != nil {
		s.Metrics.UpstreamDuration.
			WithLabelValues(cand.Provider, cand.Model).
			Observe(result.Response.Latency.Seconds())
		if usage != nil {
			s.Metrics.Tokens.WithLabelValues(tenantID, cand.Provider, cand.Model, obs.TokenKindPrompt).
				Add(float64(usage.PromptTokens))
			s.Metrics.Tokens.WithLabelValues(tenantID, cand.Provider, cand.Model, obs.TokenKindCompletion).
				Add(float64(usage.CompletionTokens))
		}
		if cost > 0 {
			s.Metrics.Cost.WithLabelValues(tenantID, cand.Provider, cand.Model).Add(cost)
		}
		s.meterCascade(tenantID, result)
	}

	rec := store.UsageRecord{
		RequestID:       requestID,
		ClientRequestID: clientRequestID,
		TenantID:        tenantID,
		KeyPrefix:       prefix,
		Provider:        cand.Provider,
		Model:           cand.Model,
		RequestedModel:  requested,
		FallbackDepth:   result.Depth,
		CostUSD:         cost,
		ChargeUSD:       charge,
		BillingMode:     s.Config.Billing.Mode,
		Streamed:        streamed,
		StatusCode:      status,
		DurationMS:      elapsed.Milliseconds(),
		RecordedAt:      time.Now().UTC(),
	}
	if usage != nil {
		rec.PromptTokens = usage.PromptTokens
		rec.CompletionToken = usage.CompletionTokens
		rec.TotalTokens = usage.TotalTokens
	}
	if s.Telemetry != nil {
		s.Telemetry.Record(rec)
	}
}

// recordRejection logs a request that never reached a provider. It writes no
// usage row: nothing was consumed, and a row with zero tokens against no model
// would distort every breakdown it appears in.
func (s *Server) recordRejection(c *fiber.Ctx, requested string, err error) {
	e := apierr.From(err)
	// The model the caller asked for still belongs on the request line, even
	// though nothing served it: "which alias is failing" is the first question
	// asked about a cascade that never landed.
	out := httpx.GetOutcome(c)
	out.Alias = requested
	httpx.SetOutcome(c, out)

	s.Logger.Warn("request rejected",
		slog.String("request_id", httpx.RequestID(c)),
		slog.String("tenant", httpx.TenantID(c)),
		slog.String("requested_model", requested),
		slog.String("code", e.Code),
		slog.Int("status", e.Status),
		slog.String("cause", causeOf(e)))
}

func causeOf(e *apierr.Error) string {
	if e.Wrapped == nil {
		return ""
	}
	return e.Wrapped.Error()
}

// timeoutAware upgrades a generic dispatch failure to 504 when the reason every
// candidate failed was that the gateway ran out of time. Reporting that as
// upstream_exhausted would send an operator looking at provider health for a
// problem that is a deadline.
//
// budget is the deadline the request was actually held to, which since GW-13 is
// the tenant's own and not necessarily the deployment's. Quoting the deployment's
// would tell a tenant narrowed to five seconds that it had two minutes.
func timeoutAware(ctx context.Context, budget time.Duration, err error) error {
	if !errors.Is(ctx.Err(), context.DeadlineExceeded) {
		return err
	}
	var e *apierr.Error
	if errors.As(err, &e) && e.Status != fiber.StatusBadGateway {
		return err
	}
	return apierr.GatewayTimeout(budget.Seconds()).WithCause(err)
}

func keyPrefix(c *fiber.Ctx) string {
	if k := httpx.Key(c); k != nil {
		return k.Prefix
	}
	return ""
}