# openapi.yaml
...
openapi: 3.1.0
info:
  title: CogniGate API
  version: 1.0.0
  summary: 'Self-hosted OpenAI-compatible gateway: routing, quotas, caching and metering across
    providers.'
  description: |
    One process serves both planes on one port.

    - **The data plane**, under `/v1`, is OpenAI-compatible. An existing SDK pointed at this base URL works unchanged, errors included.
    - **The admin plane**, under `/admin/v1`, configures tenants, keys, providers, aliases, routes, quotas and webhooks.

    A key is minted for exactly one plane and is never valid on the other; presenting a data key to `/admin/v1` fails with `wrong_plane` rather than with a 404.

    This document describes every route the gateway serves. A reconciliation test compares it against the running route table in both directions, so a route added without an entry here — or an entry here for a route that does not exist — fails the build.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
- url: http://localhost:8080
  description: The reference deployment. Both planes are on this one origin.
...
paths:
...
  /v1/chat/completions:
    post:
      tags:
      - Chat
      operationId: createChatCompletion
      summary: Create a chat completion.
      description: |
        The OpenAI-compatible entry point. `model` may be a real model id, an alias (GW-2) or the match side of a routing rule (GW-3); what actually served the request is reported in the response body and in X-CogniGate-Served-By.

        With `"stream": true` the response is `text/event-stream`. A stream that stalls terminates with an SSE error event carrying code `upstream_stream_stalled` — the HTTP status line was already sent, so it cannot be a status.
      security:
      - DataKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            example:
              model: gpt-4o-mini
              messages:
              - role: user
                content: Hello.
      responses:
        '200':
          description: A completion.
          headers:
            X-CogniGate-Request-Id:
              $ref: '#/components/headers/RequestId'
            X-CogniGate-Served-By:
              $ref: '#/components/headers/ServedBy'
            X-CogniGate-Fallback-Depth:
              $ref: '#/components/headers/FallbackDepth'
            X-CogniGate-Quota-State:
              $ref: '#/components/headers/QuotaState'
            X-CogniGate-Cache:
              $ref: '#/components/headers/Cache'
            X-CogniGate-Debug-Capture:
              $ref: '#/components/headers/DebugCapture'
            X-CogniGate-Deprecation:
              $ref: '#/components/headers/Deprecation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
            text/event-stream:
              schema:
                type: string
                description: 'Server-sent events. Each `data:` line is an OpenAI `chat.completion.chunk`,
                  and the stream ends with `data: [DONE]`.'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
...
components:
  securitySchemes:
    DataKey:
      type: http
      scheme: bearer
      description: |
        A data-plane key, prefixed `cg-`. Valid on `/v1/*` only. The prefix is part of the credential rather than decoration: it is how the gateway answers "wrong plane" without a store lookup, and how secret scanners recognise a leaked CogniGate key in a public repository.
    AdminKey:
      type: http
      scheme: bearer
      description: |
        An admin-plane key, prefixed `cga-`, or the deployment's bootstrap key. Valid on `/admin/v1/*` only. Scope is either `root`, which reaches every tenant, or `tenant:<id>`, which reaches one.
  headers:
    RequestId:
      description: |
        The identifier that ties this response to its structured log line, its usage record and, on a failure, the `request_id` in the error body. It is the one string a user can quote about a request.
      schema:
        type: string
    ServedBy:
      description: The provider and model that actually served the request.
      schema:
        type: string
    FallbackDepth:
      description: |
        How many candidates the routing cascade tried before one answered. `0` means the first choice served it.
      schema:
        type: integer
    QuotaState:
      description: |
        Where the caller stands against its GW-4 quotas after this request. `soft-exceeded` is the warning that arrives while requests are still being served.
      schema:
        type: string
        enum:
        - ok
        - soft-exceeded
        - hard-exceeded
    Cache:
      description: The GW-12 disposition of this request.
      schema:
        type: string
        enum:
        - hit
        - miss
        - bypass
    DebugCapture:
      description: Present when this request's bodies were retained under the tenant's GW-14
        capture policy.
      schema:
        type: string
    Deprecation:
      description: |
        Present when the model that served this request is marked deprecated in the catalog. GW-9 makes removing a documented header a MAJOR change, so a client may rely on this appearing.
      schema:
        type: string
    RetryAfter:
      description: Seconds to wait before retrying.
      schema:
        type: integer
  responses:
    BadRequest:
      description: The request is malformed, or a parameter is outside the range the deployment
        accepts.
      headers:
        X-CogniGate-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: No credential, an unknown one, or one minted for the other plane (`wrong_plane`).
      headers:
        X-CogniGate-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: No such resource. A tenant-scoped key naming another tenant also gets this,
        rather than a 403, so the response does not reveal whether that tenant exists.
      headers:
        X-CogniGate-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    PayloadTooLarge:
      description: The body is larger than `max_request_bytes` for this tenant.
      headers:
        X-CogniGate-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Rate limited, at the concurrency ceiling, or over a hard quota. `code`
        says which.
      headers:
        X-CogniGate-Request-Id:
          $ref: '#/components/headers/RequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
        X-CogniGate-Quota-State:
          $ref: '#/components/headers/QuotaState'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadGateway:
      description: Every candidate in the routing cascade failed. On `upstream_exhausted`
        the body enumerates them in `attempts`.
      headers:
        X-CogniGate-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServiceUnavailable:
      description: The process is draining, or a dependency it needs is unreachable.
      headers:
        X-CogniGate-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    GatewayTimeout:
      description: The upstream did not answer within `request_timeout_seconds`, or a stream
        went idle for longer than `stream_idle_timeout_seconds`.
      headers:
        X-CogniGate-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    ChatRequest:
      type: object
      required:
      - model
      - messages
      properties:
        model:
          type: string
          description: A model id, an alias, or a routing-rule match. Resolved per tenant
            (GW-1, GW-2, GW-3).
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
        stream:
          type: boolean
          description: 'When true the response is `text/event-stream`. Strictly boolean: `1`
            is refused.'
      additionalProperties: true
      description: |
        The OpenAI chat-completions request. Only the fields the gateway itself reads are described; everything else is forwarded to the provider untouched, which is what keeps an existing SDK working against this endpoint.
    ChatMessage:
      type: object
      required:
      - role
      properties:
        role:
          type: string
          enum:
          - system
          - user
          - assistant
          - tool
        content:
          type: string
      additionalProperties: true
    ChatResponse:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
          const: chat.completion
        created:
          type: integer
        model:
          type: string
          description: The model that actually served the request, which is not necessarily
            the one asked for — see X-CogniGate-Fallback-Depth.
        choices:
          type: array
          items:
            type: object
            additionalProperties: true
        usage:
          $ref: '#/components/schemas/Usage'
      additionalProperties: true
      description: |
        The provider's response, forwarded. The gateway does not rewrite the body; what it adds is carried in the X-CogniGate-* headers.
    Usage:
      type: object
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
      additionalProperties: true
...